fix: added onEXit prop to wizard
